dataplane: extract policy/zone/address-book accessors to maps_policy.go (#1686)
Same-package code motion only; method bodies byte-identical (verified).
Part of the #1686 split of pkg/dataplane/maps.go into domain accessor
files.
